// backend/src/prediction/dto/predictions.dto.ts
export class PredictionDto {
    prediction: number
}

export class DataSetPredictionDto {
    gender: string;
    age: string;
    hypertension: string;
    heart_disease: string;
    smoking_history: string;
    bmi: string;
    HbA1c_level: string;
    blood_glucose_level: string;
    diabetes: string;
}

// export class DataSetPredictionDto {
//     gender: string;
//     age: number;
//     hypertension: number;
//     heart_disease: number;
//     smoking_history: string;
//     bmi: number;
//     HbA1c_level: number;
//     blood_glucose_level: number;
//     diabetes: number;
// }